Ended working on the css
Inserted three div elements that will work as the sprites for the game. Added in styles that will change the look of the nav element by changing how it looks when the user hovers and interacts with the nav. Its background will change as well as its color and size.