Finished CSS
Finished all of the styles for the CSS stylesheet; the stylesheet is applicable to the animation. I added styles that would change the look of the links in the nav; these styles would be active when the links were hovered on /interacted with. Then, the actual "sprite" animations come up. Each sprite would have a different position on the page and a different sizes; theses were changed by the styles in the CSS. Next, I added styles that would actually be the animations for the sprites. The two "flying" sprites both shared two of the animations, the fox only had 1. These animations made it seem like the fox sprite is jumping upwards to try and catch the butterfly and bat sprite. All I need are comments for the pages.